Fix encoding issue

+++ lib/comment.rb
@@ -85,12 +85,26 @@ class Comment
   def delete(id)
     db_index.transaction do |db|
       entries = db[:entries] || []
-      entries = entries.reject do |e|
-        e['id'] == id && (@group == :super || e['user'] == @user)
-      end
+      entry = entries.find{|e| e['id'] == id}
+      raise NotFound unless entry
+      raise PermissionDenied unless @group == :super || entry['user'] == @user
+
+      entries = entries.reject {|e| e['id'] == id }
       db[:entries] = entries
 
       delete_content(id)
+    end
+  end
+
+  # Deletes all content files and the index.
+  def delete_all()
+    raise PermissionDenied unless @group == :super
+    db_index.transaction do |db|
+      entries = db[:entries] || []
+      entries.each {|e|
+        delete_content(e['id'])
+      }
+      db[:entries] = []
     end
   end
 
@@ -143,7 +157,8 @@ class Comment
 
     file = content_file(type, id)
     return unless file.exist?
-    return IO.read(file)
+    open_mode = RUBY_VERSION < '1.9.0' ? 'r' : 'r:utf-8'
+    File.open(file, open_mode) {|f| f.read }
   end
 
   def content_file(type, id)
@@ -151,13 +166,13 @@ class Comment
   end
 
   def write_content(id, content)
-    content = {
+    contents = {
       :raw  => content,
       :html => Renderer.create.render(content),
     }
-    [ :raw, :html ].each do |type|
+    contents.each do |type, content|
       open(content_file(type, id), 'w') do |io|
-        io.puts(content[type])
+        io.puts(content)
       end
     end
   end
